Move the api/types/time package to internal daemon package
