+++ daemon/internal/timestamp/timestamp.go
@@ -1,4 +1,4 @@
-package time
+package timestamp
 
 import (
 	"fmt"
